# docs/swagger-ui/openapi.yaml
{
  "components": {
    "schemas": {
...
      "AgreementRef": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "CredentialOfferDto": {
        "type": "object",
        "properties": {
          "credential_preview": {
            "$ref": "#/components/schemas/CredentialPreviewDto"
          }
        }
      },
...
      "PagedAgreementResponse": {
        "type": "object",
        "properties": {
          "pagedAgreements": {
            "$ref": "#/components/schemas/PageAgreement"
          }
        }
      },
...
      "AgreementSpecificationRef": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "TimePeriod": {
        "type": "object",
        "properties": {
          "startDateTime": {
            "type": "string"
          },
          "endDateTime": {
            "type": "string"
          }
        }
      },
...
      "Characteristic": {
        "type": "object",
        "properties": {
          "schemaLocation": {
            "type": "string"
          },
          "valueType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
      "Invitation": {
        "type": "object"
      },
      "AgreementAuthorization": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "signaturePresentation": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
...
      "RelatedParty": {
        "type": "object",
        "properties": {
          "referredType": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "extendedInfo": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "DIDStateDto": {
        "type": "object",
        "properties": {
          "updated_at": {
            "type": "string"
          },
          "credential_offer_dict": {
            "$ref": "#/components/schemas/CredentialOfferDto"
          },
          "schema_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "credential_exchange_id": {
            "type": "string"
          }
        }
      },
...
      "UpdateProductOfferingRequest": {
        "type": "object",
        "required": [
          "did",
          "geographicAddresses",
          "productOffering",
          "productOfferingPrices",
          "productSpecification",
          "resourceSpecifications",
          "serviceSpecifications"
        ],
        "properties": {
          "productOfferingPrices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOfferingPrice"
            }
          },
          "productSpecification": {
            "$ref": "#/components/schemas/ProductSpecification"
          },
          "productOffering": {
            "$ref": "#/components/schemas/ProductOffering"
          },
          "invitations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Invitation"
            },
            "type": "object"
          },
          "resourceSpecifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceSpecification"
            }
          },
          "serviceSpecifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceSpecification"
            }
          },
          "geographicAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeographicAddress"
            }
          },
          "did": {
            "type": "string"
          }
        }
      },
      "ProductOfferingPrice": {
        "type": "object",
        "properties": {
          "isBundle": {
            "type": "boolean"
          },
          "popRelationship": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOfferingPriceRelationship"
            }
          },
          "lifecycleStatus": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "recurringChargePeriodType": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "recurringChargePeriodLength": {
            "format": "int32",
            "type": "integer"
          },
          "uuid": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "percentage": {
            "format": "float",
            "type": "number"
          },
          "href": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "place": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaceRef"
            }
          },
          "bundledPopRelationship": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BundledProductOfferingPriceRelationship"
            }
          },
          "productOfferingTerm": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOfferingTerm"
            }
          },
          "prodSpecCharValueUse": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductSpecificationCharacteristicValueUse"
            }
          },
          "unitOfMeasure": {
            "$ref": "#/components/schemas/Quantity"
          },
          "validFor": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "schemaLocation": {
            "type": "string"
          },
          "priceType": {
            "type": "string"
          },
          "tax": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxItem"
            }
          },
          "version": {
            "type": "string"
          },
          "pricingLogicAlgorithm": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PricingLogicAlgorithm"
            }
          },
          "lastUpdate": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "constraint": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConstraintRef"
            }
          }
        }
      },
...
      "VerifiableCredential": {
        "type": "object"
      },
...
      "GeographicAddress": {
        "type": "object",
        "properties": {
          "country": {
            "type": "string"
          },
          "streetType": {
            "type": "string"
          },
          "geographicLocation": {
            "$ref": "#/components/schemas/GeographicLocation"
          },
          "city": {
            "type": "string"
          },
          "streetNr": {
            "type": "string"
          },
          "schemaLocation": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "streetNrLast": {
            "type": "string"
          },
          "streetNrSuffix": {
            "type": "string"
          },
          "streetName": {
            "type": "string"
          },
          "stateOrProvince": {
            "type": "string"
          },
          "streetNrLastSuffix": {
            "type": "string"
          },
          "streetSuffix": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "href": {
            "type": "string"
          }
        }
      },
...
      "PublishProductOfferingRequest": {
        "type": "object",
        "required": [
          "did",
          "geographicAddresses",
          "productOffering",
          "productOfferingPrices",
          "productSpecification",
          "resourceSpecifications",
          "serviceSpecifications"
        ],
        "properties": {
          "productOfferingPrices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOfferingPrice"
            }
          },
          "productSpecification": {
            "$ref": "#/components/schemas/ProductSpecification"
          },
          "productOffering": {
            "$ref": "#/components/schemas/ProductOffering"
          },
          "invitations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Invitation"
            },
            "type": "object"
          },
          "verifiableCredentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerifiableCredential"
            }
          },
          "resourceSpecifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceSpecification"
            }
          },
          "serviceSpecifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceSpecification"
            }
          },
          "geographicAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeographicAddress"
            }
          },
          "did": {
            "type": "string"
          }
        }
      },
...
      "AgreementItem": {
        "type": "object",
        "properties": {
          "termOrCondition": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementTermOrCondition"
            }
          },
          "product": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductRef"
            }
          },
          "productOffering": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductOfferingRef"
            }
          },
          "schemaLocation": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "baseType": {
            "type": "string"
          }
        }
      },
...
      "Agreement": {
        "type": "object",
        "properties": {
          "agreementPeriod": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "agreementType": {
            "type": "string"
          },
          "associatedAgreement": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementRef"
            }
          },
          "schemaLocation": {
            "type": "string"
          },
          "documentNumber": {
            "format": "int32",
            "type": "integer"
          },
          "agreementSpecification": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementSpecificationRef"
            }
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "initialDate": {
            "format": "date-time",
            "type": "string"
          },
          "characteristic": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Characteristic"
            }
          },
          "baseType": {
            "type": "string"
          },
          "agreementAuthorization": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementAuthorization"
            }
          },
          "name": {
            "type": "string"
          },
          "completionDate": {
            "$ref": "#/components/schemas/TimePeriod"
          },
          "statementOfIntent": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "agreementItem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreementItem"
            }
          },
          "engagedParty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedParty"
            }
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8087"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
    "/api/v1/agreement/{identifier}": {
      "get": {
        "description": "Get an agreement by id or DID",
        "operationId": "getAgreement",
        "responses": {
          "200": {
            "description": "An agreement matching the provided identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agreement"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id was provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An  couldn\u0027t be found with the provided Identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "identifier",
            "required": true
          }
        ],
        "tags": [
          "Agreement"
        ]
      }
    },
    "/api/v1/agreement/{id}/identity": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DIDStateDto"
              }
            }
          },
          "required": true
        },
        "description": "Callback endpoint to handle processing async DID identifier generation",
        "operationId": "updateTemplateIdentity_1",
        "responses": {
          "200": {
            "description": "The Agreement was updated successfully"
          },
          "400": {
            "description": "Invalid parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An Agreement couldn\u0027t be found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "Agreement"
        ]
      }
    },
...
    "/api/v1/agreement/{id}": {
      "delete": {
        "description": "Delete an Agreement definition by DID",
        "operationId": "removeAgreement",
        "responses": {
          "200": {
            "description": "The Agreement was deleted successfully"
          },
          "400": {
            "description": "Invalid parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An Agreement couldn\u0027t be found with the provided DID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "did",
            "required": true
          }
        ],
        "tags": [
          "Agreement"
        ]
      }
    },
    "/api/v1/service-level-agreement/{slaId}/identity": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DIDStateDto"
              }
            }
          },
          "required": true
        },
        "description": "Callback endpoint to handle processing async DID identifier generation",
        "operationId": "updateTemplateIdentity",
        "responses": {
          "200": {
            "description": "The SLA was updated successfully"
          },
          "400": {
            "description": "Invalid parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "An SLA couldn\u0027t be found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "slaId",
            "required": true
          }
        ],
        "tags": [
          "Service Level Agreement"
        ]
      }
    },
...
    "/product-offer/{offerId}": {
      "delete": {
        "operationId": "removeProductOffering",
        "responses": {
          "200": {
            "description": "Product offering removed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "offerId",
            "required": true
          }
        ],
        "tags": [
          "Product Offer"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductOfferingRequest"
              }
            }
          },
          "required": true
        },
        "operationId": "updateProductOffering",
        "responses": {
          "200": {
            "description": "Product offering updated",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "offerId",
            "required": true
          }
        ],
        "tags": [
          "Product Offer"
        ]
      }
    },
    "/product-offer": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishProductOfferingRequest"
              }
            }
          },
          "required": true
        },
        "operationId": "publishProductOffering",
        "responses": {
          "200": {
            "description": "Published product offering",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Product Offer"
        ]
      }
    },
    "/api/v1/agreement": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Agreement"
              }
            }
          },
          "required": true
        },
        "description": "Create a new Service Level Agreement definition",
        "operationId": "createAgreement",
        "responses": {
          "201": {
            "description": "A Agreement definition was created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Agreement definition provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Agreement"
        ]
      },
      "get": {
        "description": "Retrieve a paged collection of Agreements according to paging and filter parameters",
        "operationId": "getAgreements",
        "responses": {
          "200": {
            "description": "A Paged List of Agreements",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAgreementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid page or filter parameters provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "default": "0",
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "description": "Page you want to retrieve (0..N)"
          },
          {
            "schema": {
              "default": "20",
              "type": "integer"
            },
            "in": "query",
            "name": "size",
            "description": "Number of records per page."
          },
          {
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "in": "query",
            "name": "sort",
            "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported."
          }
        ],
        "tags": [
          "Agreement"
        ]
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}